Move follow action to UserFragment
This will allow to check if the user is an organization and disable the follow action. Otherwise we would need to load user data in 2 places.
