+++ app/src/main/java/com/gh4a/fragment/UserFragment.java
@@ -20,12 +20,16 @@ import android.os.Bundle;
 import android.support.v4.content.Loader;
 import android.text.format.DateFormat;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.gh4a.BackgroundTask;
 import com.gh4a.Gh4Application;
 import com.gh4a.R;
 import com.gh4a.activities.FollowerFollowingListActivity;
@@ -34,6 +38,7 @@ import com.gh4a.activities.OrganizationMemberListActivity;
 import com.gh4a.activities.RepositoryActivity;
 import com.gh4a.activities.RepositoryListActivity;
 import com.gh4a.activities.UserActivity;
+import com.gh4a.loader.IsFollowingUserLoader;
 import com.gh4a.loader.LoaderCallbacks;
 import com.gh4a.loader.LoaderResult;
 import com.gh4a.loader.OrganizationListLoader;
@@ -45,6 +50,7 @@ import com.gh4a.utils.StringUtils;
 
 import org.eclipse.egit.github.core.Repository;
 import org.eclipse.egit.github.core.User;
+import org.eclipse.egit.github.core.service.UserService;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -65,6 +71,8 @@ public class UserFragment extends LoadingFragmentBase implements View.OnClickLis
     private String mUserLogin;
     private User mUser;
     private View mContentView;
+    private Boolean mIsFollowing;
+    private boolean mIsSelf;
 
     private final LoaderCallbacks<User> mUserCallback = new LoaderCallbacks<User>(this) {
         @Override
@@ -107,10 +115,24 @@ public class UserFragment extends LoadingFragmentBase implements View.OnClickLis
         }
     };
 
+    private final LoaderCallbacks<Boolean> mIsFollowingCallback = new LoaderCallbacks<Boolean>(this) {
+        @Override
+        protected Loader<LoaderResult<Boolean>> onCreateLoader() {
+            return new IsFollowingUserLoader(getActivity(), mUserLogin);
+        }
+        @Override
+        protected void onResultReady(Boolean result) {
+            mIsFollowing = result;
+            getActivity().invalidateOptionsMenu();
+        }
+    };
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         mUserLogin = getArguments().getString("login");
+        mIsSelf = ApiHelpers.loginEquals(mUserLogin, Gh4Application.get().getAuthLogin());
+        setHasOptionsMenu(true);
     }
 
     @Override
@@ -122,11 +144,13 @@ public class UserFragment extends LoadingFragmentBase implements View.OnClickLis
     @Override
     public void onRefresh() {
         mUser = null;
+        mIsFollowing = false;
         if (mContentView != null) {
             fillOrganizations(null);
             fillTopRepos(null);
         }
-        hideContentAndRestartLoaders(0);
+        hideContentAndRestartLoaders(0, 3);
+        getActivity().invalidateOptionsMenu();
     }
 
     @Override
@@ -136,6 +160,49 @@ public class UserFragment extends LoadingFragmentBase implements View.OnClickLis
         setContentShown(false);
 
         getLoaderManager().initLoader(0, null, mUserCallback);
+
+        if (!mIsSelf && Gh4Application.get().isAuthorized()) {
+            getLoaderManager().initLoader(3, null, mIsFollowingCallback);
+        }
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.user_follow_menu, menu);
+        super.onCreateOptionsMenu(menu, inflater);
+    }
+
+    @Override
+    public void onPrepareOptionsMenu(Menu menu) {
+        super.onPrepareOptionsMenu(menu);
+
+        MenuItem followAction = menu.findItem(R.id.follow);
+        if (followAction != null) {
+            if (!mIsSelf && Gh4Application.get().isAuthorized()) {
+                followAction.setVisible(true);
+                if (mIsFollowing == null) {
+                    followAction.setActionView(R.layout.ab_loading);
+                    followAction.expandActionView();
+                } else if (mIsFollowing) {
+                    followAction.setTitle(R.string.user_unfollow_action);
+                } else {
+                    followAction.setTitle(R.string.user_follow_action);
+                }
+            } else {
+                followAction.setVisible(false);
+            }
+        }
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.follow) {
+            item.setActionView(R.layout.ab_loading);
+            item.expandActionView();
+            new UpdateFollowTask().schedule();
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
     }
 
     private void fillData() {
@@ -325,17 +392,42 @@ public class UserFragment extends LoadingFragmentBase implements View.OnClickLis
         }
     }
 
-    public void updateFollowingAction(boolean following) {
+    public void updateFollowingAction() {
         if (mUser == null) {
             return;
         }
 
-        if (following) {
+        if (mIsFollowing) {
             mUser.setFollowers(mUser.getFollowers() + 1);
         } else {
             mUser.setFollowers(mUser.getFollowers() - 1);
         }
         TextView tvFollowersCount = mContentView.findViewById(R.id.tv_followers_count);
         tvFollowersCount.setText(String.valueOf(mUser.getFollowers()));
+    }
+
+    private class UpdateFollowTask extends BackgroundTask<Void> {
+        public UpdateFollowTask() {
+            super(UserFragment.this.getActivity());
+        }
+
+        @Override
+        protected Void run() throws Exception {
+            UserService userService = (UserService)
+                    Gh4Application.get().getService(Gh4Application.USER_SERVICE);
+            if (mIsFollowing) {
+                userService.unfollow(mUserLogin);
+            } else {
+                userService.follow(mUserLogin);
+            }
+            return null;
+        }
+
+        @Override
+        protected void onSuccess(Void result) {
+            mIsFollowing = !mIsFollowing;
+            updateFollowingAction();
+            getActivity().invalidateOptionsMenu();
+        }
     }
 }
