+++ app/src/main/res/menu/user_follow_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/follow"
+        android:title="@string/user_follow_action"
+        app:showAsAction="always" />
+
+</menu>
