+++ app/src/main/res/menu/user_menu.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/follow"
-        android:title="@string/user_follow_action"
-        app:showAsAction="always" />
     <item android:id="@+id/share"
         android:title="@string/share"
         app:showAsAction="never" />
